Start to generate WHERE SQL string from Linq expression
