+++ src/Linq2CouchBaseLiteExpression.Tests/Linq2CouchBaseLiteExpression.Tests.csproj
@@ -1,21 +1,21 @@
-﻿<Project Sdk="MSBuild.Sdk.Extras">
+﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>netcoreapp3.1</TargetFramework>
+    <TargetFramework>net6.0</TargetFramework>
     <ProjectGuid>{49A4AD2F-5F22-57A4-8436-9FBA3C6D72DE}</ProjectGuid>
     <IsPackable>false</IsPackable>
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Microsoft.NET.Test.Sdk" Version="16.11.0" />
-    <PackageReference Include="MSTest.TestAdapter" Version="2.2.7" />
-    <PackageReference Include="MSTest.TestFramework" Version="2.2.7" />
-    <PackageReference Include="coverlet.collector" Version="3.1.0">
+    <PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.1.0" />
+    <PackageReference Include="MSTest.TestAdapter" Version="2.2.8" />
+    <PackageReference Include="MSTest.TestFramework" Version="2.2.8" />
+    <PackageReference Include="coverlet.collector" Version="3.1.2">
       <PrivateAssets>all</PrivateAssets>
       <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
     </PackageReference>
-    <PackageReference Include="Faker.Net" Version="1.5.148" />
-    <PackageReference Include="Couchbase.Lite.Support.NetDesktop" Version="2.8.6" />
+    <PackageReference Include="Faker.Net" Version="2.0.154" />
+    <PackageReference Include="Couchbase.Lite.Support.NetDesktop" Version="3.0.0" />
   </ItemGroup>
 
   <ItemGroup>
