--- src/global.json
@@ -1,5 +1,0 @@
-{
-    "msbuild-sdks": {
-        "MSBuild.Sdk.Extras": "3.0.23"
-    }
-}
